speedtest-svc/requirements.txt
Flask==2.0.1
flask_restful==0.3.9
speedtest-cli==2.1.3
